Convert wind
