+++ tests/solar/test_solar.py
@@ -8,10 +8,6 @@
 #
 
 
-"""
-Tests for the high level solar functions
-"""
-
 import datetime
 
 import numpy as np
